feat(document-grid): add drag'n'drop capability to document grid
This duplicates quite a few lines of code. But I didn't see a way
to reuse the code from the upload button component. Maybe we should
move those actions to a service that can be used by all components
where we could also put the upload and delete actions.
But this is - IMO - out of scope for this PR. I will gladly make
another PR with the service.