fix(deps): remove direct dependency to ember-data
Since `ember-engines` v0.11 engines will now be properly isolated from
the app which means that no classes like adapters etc. will be leaked
into the host app. In order to handle this, the host app will need to
define the serializers, adapters and models for `ember-alexandria` in
their host application. This is motivation enough to decouple the hard
dependency to `ember-data` which resulted in an own store inside our
engine. In order to achieve that, `ember-alexandria` needs to present a
solution for its custom adapters: Host applications need to be able to
inject authorization mechanisms into the engine. This is achieved by
`ember-alexandria` exporting factories for the adapters instead of
classes. This way, host applications can call the factory with their
base class that handles authentication.

BREAKING CHANGE: Please refer to the [documentation](MIGRATIONS.md).